refactor

+++ library/src/main/scala/com/github/j5ik2o/bacs/model/BoardState.scala
@@ -1,6 +1,6 @@
 package com.github.j5ik2o.bacs.model
 
-case class BoardStateData(special_quotation: Long)
+case class BoardStateData(special_quotation: BigDecimal)
 case class BoardState(health: String,
                       state: String,
                       data: Option[BoardStateData])
