+++ library/src/main/scala/com/github/j5ik2o/bacs/model/CoinIn.scala
@@ -3,7 +3,7 @@ package com.github.j5ik2o.bacs.model
 case class CoinIn(id: Long,
                   order_id: String,
                   currency_code: String,
-                  amount: Double,
+                  amount: BigDecimal,
                   address: String,
                   tx_hash: String,
                   status: String,
