+++ library/src/main/scala/com/github/j5ik2o/bacs/model/CoinOut.scala
@@ -3,10 +3,10 @@ package com.github.j5ik2o.bacs.model
 case class CoinOut(id: Long,
                    order_id: String,
                    currency_code: String,
-                   amount: Double,
+                   amount: BigDecimal,
                    address: String,
                    tx_hash: String,
-                   fee: Double,
-                   additional_fee: Double,
+                   fee: BigDecimal,
+                   additional_fee: BigDecimal,
                    status: String,
                    event_date: String)
