+++ library/src/main/scala/com/github/j5ik2o/bacs/model/Collateral.scala
@@ -1,6 +1,6 @@
 package com.github.j5ik2o.bacs.model
 
-case class Collateral(collateral: Double,
-                      open_position_pnl: Double,
-                      require_collateral: Double,
-                      keep_rate: Double)
+case class Collateral(collateral: BigDecimal,
+                      open_position_pnl: BigDecimal,
+                      require_collateral: BigDecimal,
+                      keep_rate: BigDecimal)
