+++ library/src/main/scala/com/github/j5ik2o/bacs/model/Deposit.scala
@@ -3,6 +3,6 @@ package com.github.j5ik2o.bacs.model
 case class Deposit(id: Long,
                    order_id: String,
                    currency_code: String,
-                   amount: Double,
+                   amount: BigDecimal,
                    status: String,
                    event_date: String)
